Extract tool dropdown component
